// src/types.ts
    export interface Artwork {
    id: number
    title: string
    place_of_origin: string | null
    artist_display: string | null
    inscriptions: string | null
    date_start: number | null
    date_end: number | null
    }

    export interface ApiPagination {
    total: number
    limit: number
    offset: number
    total_pages: number
    current_page: number
    }

    export interface ApiResponse {
    data: Artwork[]
    pagination: ApiPagination
    }